% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press.R
\name{temp_init}
\alias{temp_init}
\title{Pressure Calculations}
\usage{
temp_init(est = NULL, type = "tdb", units = "celsius")
}
\arguments{
\item{est}{a numeric value representing the measurement of interest.}

\item{type}{the type of pressure calculation.}

\item{units}{the units of measurement. Options include torr, atm, mb (millibar), and psi. Default is torr.}
}
\description{
Functions for calculating different forms of pressure
}
\section{Functions}{
\itemize{
\item \code{temp_init}: Pressure measurement
}}

