% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_functions.R
\name{temp_convert}
\alias{temp_convert}
\alias{press_convert}
\title{Unit Conversion Functions}
\usage{
temp_convert(
  x = NULL,
  to = thermoreg_options("temp_scale"),
  from = "fahrenheit",
  type = "tdb"
)

press_convert(
  x = NULL,
  to = thermoreg_options("press_scale"),
  from = "mbar",
  type = "barometric"
)
}
\arguments{
\item{x}{a numeric value representing the measurement of interest}

\item{to}{the scale to convert to}

\item{from}{the scale to convert from}

\item{type}{scale of measurement (i.e., tdb)}
}
\description{
Functions for providing conversions between scales of measurement.
}
\section{Functions}{
\itemize{
\item \code{temp_convert}: Conversion for Temperature

\item \code{press_convert}: Conversion for Pressure
}}

