% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press.R
\name{press_init}
\alias{press_init}
\title{Pressure Calculations}
\usage{
press_init(
  est = NULL,
  type = "barometric",
  units = thermoreg_options("press_scale")
)
}
\arguments{
\item{est}{a numeric value representing the measurement of interest.}

\item{type}{the type of pressure calculation.}

\item{units}{the units of measurement. Options include torr, atm, mbar (millibar), and psi. Default is torr.}
}
\description{
Functions for calculating different forms of pressure
}
\section{Functions}{
\itemize{
\item \code{press_init}: Pressure measurement
}}

