% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{temp_init}
\alias{temp_init}
\title{Temperature Calculations}
\usage{
temp_init(est = NULL, type = "tdb", units = "celsius")
}
\arguments{
\item{est}{a numeric value representing the measurement of interest.}

\item{type}{the type of temperature calculation.}

\item{units}{the units of measurement. Options include Celsius, Fahrenheit, or Kelvin. Default is celsius.}
}
\description{
Functions for calculating different forms of temperature.
}
\section{Functions}{
\itemize{
\item \code{temp_init}: Temperature measurement
}}

